Move UL $ur/sr command from pilart to unison cog
