+++ cogs/unison.py
@@ -7,13 +7,17 @@ implement UTC timing with as little effort as possible. :<
 from asyncio import sleep
 from configparser import ConfigParser
 from datetime import datetime
+from io import BytesIO
 from os import listdir, mkdir
 from os.path import isdir, isfile, join
 
+import aiohttp
+from PIL import Image
+from discord import File
 from discord.ext import commands
 from discord.ext.commands import BucketType
 
-from nyxutils import list_string, reply
+from nyxutils import get_member, list_string, reply
 
 folder = "unison"
 events_folder = "events"
@@ -31,6 +35,8 @@ reminders = {}
 days = ["mon", "tue", "wed", "thu", "fri", "sat", "sun"]
 days_full = ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday",
              "Saturday", "Sunday"]
+elements = ["fire", "water", "wind", "light", "dark", "haste"]
+times_list_threshold = 10
 
 
 class Event:
@@ -232,8 +238,11 @@ def is_day(string):
 
 
 def is_event_code(string, force_enable=False):
-    enabled = not events[a].disabled or force_enable
-    return any(string == a.lower() and enabled for a in events)
+    def enabled(a):
+        return events[a].disabled or force_enable
+
+    # enabled = not events[a].disabled or force_enable
+    return any(string == a.lower() and enabled(a) for a in events)
 
 
 def time_stamp(datetime):
@@ -494,7 +503,8 @@ def save_reminders(uid, debug_name=None):
     user_file.close()
 
 
-times_list_threshold = 10
+def get_asset(file_name):
+    return join(folder, file_name)
 
 
 class Unison:
@@ -786,6 +796,68 @@ class Unison:
                     results.append(list_string(times_list, key=key))
             results = "".join(results)
             await reply(ctx, results)
+
+    @commands.command(aliases=["sr"])
+    @commands.bot_has_permissions(send_messages=True, attach_files=True)
+    @commands.cooldown(1, 5, BucketType.user)
+    async def ur(self, ctx, element: str, user: str = None):
+        """Creates a UR monster of yourself or another user.
+        Credit to Bevgebra for the templates...
+        """
+        ele = None
+        element = element.lower()
+        for thing in elements:
+            if thing in element.lower():
+                ele = thing
+                break
+        if ele is None:
+            await reply(ctx, "Invalid element!")
+            ctx.command.reset_cooldown(ctx)
+            return
+        else:
+            element = ele[:1].upper() + ele[1:]
+        if user is not None:
+            user = await get_member(ctx, user)
+            if user is None:
+                await reply(ctx, "I dun know who you are talking about...")
+                ctx.command.reset_cooldown(ctx)
+                return
+        else:
+            user = ctx.message.author
+        rarity = ctx.invoked_with.upper()
+        url = user.avatar_url
+        if not url:
+            url = user.default_avatar_url
+        print(url)
+
+        async with ctx.message.channel.typing(), aiohttp.ClientSession(
+                loop=self.nyx.loop) as session, session.get(url) as req:
+            if req.status == 200:
+                imfile = BytesIO(await req.read())
+                img = Image.open(imfile)
+                snum = 5
+                ednum = snum + 426
+                base = Image.open(get_asset(rarity + "Base.png"))
+                over = Image.open(
+                    get_asset(rarity + "Overlay" + element + ".png"))
+                img = img.resize((426, 426), Image.LANCZOS)
+                mask = img if "RGBA" in img.mode else None
+                base.paste(img, (snum, snum, ednum, ednum), mask=mask)
+                base.paste(over, mask=over)
+
+                # Save in-memory filestream and send to Discord
+                imfile = BytesIO()
+                base.save(imfile, format="png")
+                # Move pointer to beginning so Discord can read pic.
+                imfile.seek(0)
+                msg = "New monster released in Unison League!"
+                if ctx.guild is not None:
+                    msg = ctx.message.author.mention + ", n" + msg[1:]
+                await ctx.send(msg,
+                               file=File(imfile, filename=rarity + ".png"))
+                imfile.close()
+            else:
+                await reply(ctx, "Image loading failed! :<")
 
     async def clock(self):
         await self.nyx.wait_until_ready()
